cmake-utils.eclass: use a proper if statement

+++ eclass/cmake-utils.eclass
@@ -404,12 +404,12 @@ _modify-cmakelists() {
 _cleanup_cmake() {
 	: ${CMAKE_USE_DIR:=${S}}
 
-	[[ "${CMAKE_REMOVE_MODULES}" == "yes" ]] && {
+	if [[ "${CMAKE_REMOVE_MODULES}" == "yes" ]] ; then
 		local name
 		for name in ${CMAKE_REMOVE_MODULES_LIST} ; do
 			find "${S}" -name ${name}.cmake -exec rm -v {} + || die
 		done
-	}
+	fi
 
 	# check if CMakeLists.txt exist and if no then die
 	if [[ ! -e ${CMAKE_USE_DIR}/CMakeLists.txt ]] ; then
